+++ app/models/list_option.rb
@@ -1,7 +1,9 @@
 class ListOption < ActiveRecord::Base
 
   # Model Relationships
-  belongs_to :option, class_name: 'StratificationFactorOption', foreign_key: 'option_id'
+  belongs_to :project
+  belongs_to :randomization_scheme
   belongs_to :list
+  belongs_to :option, class_name: 'StratificationFactorOption', foreign_key: 'option_id'
 
 end