+++ test/controllers/block_size_multipliers_controller_test.rb
@@ -4,8 +4,12 @@ class BlockSizeMultipliersControllerTest < ActionController::TestCase
   setup do
     login(users(:valid))
     @project = projects(:one)
-    @randomization_scheme = randomization_schemes(:one)
-    @block_size_multiplier = block_size_multipliers(:one)
+
+    @published_scheme = randomization_schemes(:one)
+    @published_block_size_multiplier = block_size_multipliers(:one)
+
+    @randomization_scheme = randomization_schemes(:two)
+    @block_size_multiplier = block_size_multipliers(:six)
   end
 
   test "should get index" do
@@ -19,12 +23,25 @@ class BlockSizeMultipliersControllerTest < ActionController::TestCase
     assert_response :success
   end
 
+  test "should not get new for published randomization scheme" do
+    get :new, project_id: @project, randomization_scheme_id: @published_scheme
+    assert_redirected_to project_randomization_scheme_block_size_multipliers_path(assigns(:project), assigns(:randomization_scheme))
+  end
+
   test "should create block_size_multiplier" do
     assert_difference('BlockSizeMultiplier.count') do
       post :create, project_id: @project, randomization_scheme_id: @randomization_scheme, block_size_multiplier: { value: 5, allocation: 1 }
     end
 
     assert_redirected_to project_randomization_scheme_block_size_multiplier_path(assigns(:project), assigns(:randomization_scheme), assigns(:block_size_multiplier))
+  end
+
+  test "should not create block size multiplier for published randomization scheme" do
+    assert_difference('BlockSizeMultiplier.count', 0) do
+      post :create, project_id: @project, randomization_scheme_id: @published_scheme, block_size_multiplier: { value: 5, allocation: 1 }
+    end
+
+    assert_redirected_to project_randomization_scheme_block_size_multipliers_path(assigns(:project), assigns(:randomization_scheme))
   end
 
   test "should show block_size_multiplier" do
@@ -37,8 +54,20 @@ class BlockSizeMultipliersControllerTest < ActionController::TestCase
     assert_response :success
   end
 
+  test "should not get edit for published randomization scheme" do
+    get :edit, project_id: @project, randomization_scheme_id: @published_scheme, id: @published_block_size_multiplier
+    assert_redirected_to project_randomization_scheme_block_size_multiplier_path(assigns(:project), assigns(:randomization_scheme), assigns(:block_size_multiplier))
+  end
+
   test "should update block_size_multiplier" do
     patch :update, project_id: @project, randomization_scheme_id: @randomization_scheme, id: @block_size_multiplier, block_size_multiplier: { value: @block_size_multiplier.value, allocation: 3 }
+    assert_redirected_to project_randomization_scheme_block_size_multiplier_path(assigns(:project), assigns(:randomization_scheme), assigns(:block_size_multiplier))
+  end
+
+  test "should not update block size multiplier for published randomization scheme" do
+    patch :update, project_id: @project, randomization_scheme_id: @published_scheme, id: @published_block_size_multiplier, block_size_multiplier: { value: 2, allocation: 3 }
+    assert_equal 1, assigns(:block_size_multiplier).value
+    assert_equal 2, assigns(:block_size_multiplier).allocation
     assert_redirected_to project_randomization_scheme_block_size_multiplier_path(assigns(:project), assigns(:randomization_scheme), assigns(:block_size_multiplier))
   end
 
@@ -48,5 +77,13 @@ class BlockSizeMultipliersControllerTest < ActionController::TestCase
     end
 
     assert_redirected_to project_randomization_scheme_block_size_multipliers_path(assigns(:project), assigns(:randomization_scheme))
+  end
+
+  test "should not destroy block size multiplier for published randomization scheme" do
+    assert_difference('BlockSizeMultiplier.current.count', 0) do
+      delete :destroy, project_id: @project, randomization_scheme_id: @published_scheme, id: @published_block_size_multiplier
+    end
+
+    assert_redirected_to project_randomization_scheme_block_size_multiplier_path(assigns(:project), assigns(:randomization_scheme), assigns(:block_size_multiplier))
   end
 end