+++ test/controllers/randomization_schemes_controller_test.rb
@@ -7,6 +7,96 @@ class RandomizationSchemesControllerTest < ActionController::TestCase
     @randomization_scheme = randomization_schemes(:one)
   end
 
+  test "should get randomize subject for published scheme" do
+    get :randomize_subject, project_id: @project, id: @randomization_scheme
+    assert_not_nil assigns(:randomization_scheme)
+    assert_not_nil assigns(:randomization)
+    assert_response :success
+  end
+
+  test "should not get randomize subject for draft scheme" do
+    get :randomize_subject, project_id: @project, id: randomization_schemes(:two)
+    assert_not_nil assigns(:project)
+    assert_nil assigns(:randomization_scheme)
+    assert_redirected_to project_randomization_schemes_path(assigns(:project))
+  end
+
+  test "should get randomize subject for published scheme with no lists" do
+    post :randomize_subject_to_list, project_id: projects(:two), id: randomization_schemes(:three)
+    assert_not_nil assigns(:project)
+    assert_not_nil assigns(:randomization_scheme)
+    assert_response :success
+  end
+
+  test "should randomize subject for published scheme to list" do
+    # Stratification Factors { "Gender" => "Male", "Age" => "< 40" }
+    post :randomize_subject_to_list, project_id: @project, id: @randomization_scheme, subject_code: "Code02", stratification_factors: { "#{ActiveRecord::FixtureSet.identify(:gender)}" => "#{ActiveRecord::FixtureSet.identify(:male)}", "#{ActiveRecord::FixtureSet.identify(:age)}" => "#{ActiveRecord::FixtureSet.identify(:ltforty)}" }, attested: "1"
+    assert_not_nil assigns(:randomization_scheme)
+    assert_not_nil assigns(:randomization)
+    assert_redirected_to [assigns(:project), assigns(:randomization)]
+  end
+
+  test "should not randomize subject to list if already randomized" do
+    post :randomize_subject_to_list, project_id: @project, id: @randomization_scheme, subject_code: "Code01", stratification_factors: { "#{ActiveRecord::FixtureSet.identify(:gender)}" => "#{ActiveRecord::FixtureSet.identify(:male)}", "#{ActiveRecord::FixtureSet.identify(:age)}" => "#{ActiveRecord::FixtureSet.identify(:ltforty)}" }, attested: "1"
+    assert_not_nil assigns(:randomization_scheme)
+    assert_not_nil assigns(:randomization)
+    assert assigns(:randomization).errors.size > 0
+    assert_equal ["has already been randomized"], assigns(:randomization).errors[:subject_id]
+    assert_response :success
+  end
+
+  test "should not randomize subject to list if subject code is blank" do
+    post :randomize_subject_to_list, project_id: @project, id: @randomization_scheme, subject_code: "", stratification_factors: { "#{ActiveRecord::FixtureSet.identify(:gender)}" => "#{ActiveRecord::FixtureSet.identify(:male)}", "#{ActiveRecord::FixtureSet.identify(:age)}" => "#{ActiveRecord::FixtureSet.identify(:ltforty)}" }, attested: "1"
+    assert_not_nil assigns(:randomization_scheme)
+    assert_not_nil assigns(:randomization)
+    assert assigns(:randomization).errors.size > 0
+    assert_equal ["can't be blank"], assigns(:randomization).errors[:subject_code]
+    assert_response :success
+  end
+
+  test "should not randomize subject to list if missing one or more stratification factors" do
+    post :randomize_subject_to_list, project_id: @project, id: @randomization_scheme, subject_code: "Code02", stratification_factors: { "#{ActiveRecord::FixtureSet.identify(:gender)}" => "#{ActiveRecord::FixtureSet.identify(:male)}" }, attested: "1"
+    assert_not_nil assigns(:randomization_scheme)
+    assert_not_nil assigns(:randomization)
+    assert assigns(:randomization).errors.size > 0
+    assert_equal ["can't be blank"], assigns(:randomization).errors[:stratification_factors]
+    assert_response :success
+  end
+
+  test "should not randomize subject to list if missing all stratification factors" do
+    post :randomize_subject_to_list, project_id: @project, id: @randomization_scheme, subject_code: "Code02", attested: "1"
+    assert_not_nil assigns(:randomization_scheme)
+    assert_not_nil assigns(:randomization)
+    assert assigns(:randomization).errors.size > 0
+    assert_equal ["can't be blank"], assigns(:randomization).errors[:stratification_factors]
+    assert_response :success
+  end
+
+  test "should not randomize subject to list if attestation is not checked" do
+    post :randomize_subject_to_list, project_id: @project, id: @randomization_scheme, subject_code: "Code02", stratification_factors: { "#{ActiveRecord::FixtureSet.identify(:gender)}" => "#{ActiveRecord::FixtureSet.identify(:male)}", "#{ActiveRecord::FixtureSet.identify(:age)}" => "#{ActiveRecord::FixtureSet.identify(:ltforty)}" }, attested: "0"
+    assert_not_nil assigns(:randomization_scheme)
+    assert_not_nil assigns(:randomization)
+    assert assigns(:randomization).errors.size > 0
+    assert_equal ["must be checked"], assigns(:randomization).errors[:attested]
+    assert_response :success
+  end
+
+  test "should not randomize subject if no lists have been generated for randomization scheme" do
+    post :randomize_subject_to_list, project_id: projects(:two), id: randomization_schemes(:three), subject_code: "2TWO02", stratification_factors: { }, attested: "1"
+    assert_not_nil assigns(:randomization_scheme)
+    assert_not_nil assigns(:randomization)
+    assert assigns(:randomization).errors.size > 0
+    assert_equal ["need to be generated before a subject can be randomized"], assigns(:randomization).errors[:lists]
+    assert_response :success
+  end
+
+  test "should not randomize subject to list for draft scheme" do
+    post :randomize_subject_to_list, project_id: @project, id: randomization_schemes(:two)
+    assert_not_nil assigns(:project)
+    assert_nil assigns(:randomization_scheme)
+    assert_redirected_to project_randomization_schemes_path(assigns(:project))
+  end
+
   test "should get index" do
     get :index, project_id: @project
     assert_response :success
@@ -26,6 +116,17 @@ class RandomizationSchemesControllerTest < ActionController::TestCase
     assert_redirected_to project_randomization_scheme_path(assigns(:project), assigns(:randomization_scheme))
   end
 
+  test "should not create randomization scheme with blank name" do
+    assert_difference('RandomizationScheme.count', 0) do
+      post :create, project_id: @project, randomization_scheme: { name: "", description: @randomization_scheme.description, published: @randomization_scheme.published, randomization_goal: @randomization_scheme.randomization_goal }
+    end
+    assert_not_nil assigns(:randomization_scheme)
+    assert assigns(:randomization_scheme).errors.size > 0
+    assert_equal ["can't be blank"], assigns(:randomization_scheme).errors[:name]
+    assert_template 'new'
+    assert_response :success
+  end
+
   test "should show randomization_scheme" do
     get :show, project_id: @project, id: @randomization_scheme
     assert_response :success
@@ -39,6 +140,15 @@ class RandomizationSchemesControllerTest < ActionController::TestCase
   test "should update randomization_scheme" do
     patch :update, project_id: @project, id: @randomization_scheme, randomization_scheme: { name: "Updated Randomization Scheme", description: @randomization_scheme.description, published: @randomization_scheme.published, randomization_goal: @randomization_scheme.randomization_goal }
     assert_redirected_to project_randomization_scheme_path(assigns(:project), assigns(:randomization_scheme))
+  end
+
+  test "should not update randomization scheme with existing name" do
+    patch :update, project_id: @project, id: @randomization_scheme, randomization_scheme: { name: "Randomization Scheme 2", description: @randomization_scheme.description, published: @randomization_scheme.published, randomization_goal: @randomization_scheme.randomization_goal }
+    assert_not_nil assigns(:randomization_scheme)
+    assert assigns(:randomization_scheme).errors.size > 0
+    assert_equal ["has already been taken"], assigns(:randomization_scheme).errors[:name]
+    assert_template 'edit'
+    assert_response :success
   end
 
   test "should destroy randomization_scheme" do